chore: added hidden --outdate option for cache clear

+++ src/cli/cache/clear.rs
@@ -1,7 +1,8 @@
-use eyre::Result;
-
 use crate::dirs::CACHE;
 use crate::file::{display_path, remove_all};
+use eyre::Result;
+use filetime::set_file_times;
+use walkdir::WalkDir;
 
 /// Deletes all cache files in mise
 #[derive(Debug, clap::Args)]
@@ -10,6 +11,10 @@ pub struct CacheClear {
     /// Plugin(s) to clear cache for
     /// e.g.: node, python
     plugin: Option<Vec<String>>,
+
+    /// Mark all cache files as old
+    #[clap(long, hide = true)]
+    outdate: bool,
 }
 
 impl CacheClear {
@@ -18,15 +23,34 @@ impl CacheClear {
             Some(plugins) => plugins.iter().map(|p| CACHE.join(p)).collect(),
             None => vec![CACHE.to_path_buf()],
         };
-        for p in cache_dirs {
-            if p.exists() {
-                debug!("clearing cache from {}", display_path(&p));
-                remove_all(p)?;
+        if self.outdate {
+            for p in cache_dirs {
+                if p.exists() {
+                    debug!("outdating cache from {}", display_path(&p));
+                    let files = WalkDir::new(&p)
+                        .into_iter()
+                        .filter_map(|e| e.ok())
+                        .filter(|e| e.file_type().is_file() || e.file_type().is_dir());
+                    for e in files {
+                        set_file_times(
+                            e.path(),
+                            filetime::FileTime::zero(),
+                            filetime::FileTime::zero(),
+                        )?;
+                    }
+                }
             }
-        }
-        match &self.plugin {
-            Some(plugins) => info!("cache cleared for {}", plugins.join(", ")),
-            None => info!("cache cleared"),
+        } else {
+            for p in cache_dirs {
+                if p.exists() {
+                    debug!("clearing cache from {}", display_path(&p));
+                    remove_all(p)?;
+                }
+            }
+            match &self.plugin {
+                Some(plugins) => info!("cache cleared for {}", plugins.join(", ")),
+                None => info!("cache cleared"),
+            }
         }
         Ok(())
     }
